fix: make a lot of crates optional
Saves a lot of compile time and energy!
