+++ Cargo.toml
@@ -13,10 +13,11 @@ exclude = [".github/"]
 
 [features]
 default = ["backend-kvm", "backend-sev", "backend-sgx", "backend-nil"]
-backend-kvm = ["kvm-bindings", "kvm-ioctls", "x86_64"]
-backend-sev = ["backend-kvm", "ureq"]
-backend-sgx = ["sgx", "ureq", "x86_64"]
-backend-nil = ["enarx-exec-wasmtime"]
+backend-kvm = ["dep:const-default", "dep:iocuddle", "dep:kvm-bindings", "dep:kvm-ioctls", "dep:lset", "dep:rand", "dep:x86_64", "load-binary"]
+backend-sev = ["dep:ureq", "dep:nbytes", "backend-kvm"]
+backend-sgx = ["dep:iocuddle", "dep:protobuf", "dep:sgx", "dep:ureq", "dep:vdso", "dep:x86_64", "load-binary"]
+backend-nil = ["dep:enarx-exec-wasmtime"]
+load-binary = ["dep:goblin", "dep:mmarinus", "dep:primordial", "dep:sallyport", "dep:semver", "dep:static_assertions"]
 
 # non-default features
 gdb = ["gdbstub", "enarx-shim-kvm/gdb", "enarx-shim-sgx/gdb"]
@@ -29,35 +30,35 @@ atty = { version = "0.2", default-features = false }
 bitflags = { version = "1.2", default-features = false }
 clap = { version = "3.1", features = ["env", "derive", "std"], default-features = false }
 colorful = { version = "0.2", default-features = false }
-const-default = { version = "1.0", features = ["derive"], default-features = false }
 dirs = { version = "4.0", default-features = false }
 env_logger = { version = "0.9", default-features = false }
-goblin = { version = "0.5", features = ["elf64"], default-features = false }
-iocuddle = { version = "0.1.1", default-features = false }
 libc = { version = "0.2", default-features = false }
 log = { version = "0.4", default-features = false }
-lset = { version = "0.3", default-features = false }
-mmarinus = { version = "0.4.0", default-features = false }
-nbytes = { version = "0.1", default-features = false }
 once_cell = { version = "1.10.0", features = ["std"], default-features = false }
-primordial = { version = "0.5", features = ["alloc"], default-features = false }
-protobuf = { version = "2.22", default-features = false }
-rand = { version = "0.8", features = ["std", "std_rng"], default-features = false }
-sallyport = { version = "0.4.0", path = "crates/sallyport", default-features = false }
-semver = { version = "1.0", default-features = false }
 serde = { version = "1.0.136", features = ["derive"], default-features = false }
 serde_json = { version = "1.0.79", features = ["std"], default-features = false }
-static_assertions = { version = "1.1.0", default-features = false }
-vdso = { version = "0.2", default-features = false }
 
 # optional dependencies
+const-default = { version = "1.0", optional = true, features = ["derive"], default-features = false }
+enarx-exec-wasmtime = { version = "0.4.0", path = "crates/exec-wasmtime", optional = true, default-features = false }
 gdbstub = { version = "0.5.0", optional = true, features = ["std"], default-features = false }
+goblin = { version = "0.5", optional = true, features = ["elf64", "elf32", "std", "endian_fd"], default-features = false }
+iocuddle = { version = "0.1.1", optional = true, default-features = false }
 kvm-bindings = { version = "0.5", optional = true, default-features = false }
 kvm-ioctls = { version = "0.11", optional = true, default-features = false }
+lset = { version = "0.3", optional = true, default-features = false }
+mmarinus = { version = "0.4.0", optional = true, default-features = false }
+nbytes = { version = "0.1", optional = true, default-features = false }
+primordial = { version = "0.5", optional = true, features = ["alloc"], default-features = false }
+protobuf = { version = "2.22", optional = true, default-features = false }
+rand = { version = "0.8", optional = true, features = ["std", "std_rng"], default-features = false }
+sallyport = { version = "0.4.0", optional = true, path = "crates/sallyport", default-features = false }
+semver = { version = "1.0", optional = true, default-features = false }
 sgx = { version = "0.4.1", optional = true, features = ["rcrypto"], default-features = false }
+static_assertions = { version = "1.1.0", optional = true, default-features = false }
 ureq = { version = "2.4.0", optional = true, default-features = false }
+vdso = { version = "0.2", optional = true, default-features = false }
 x86_64 = { version = "0.14.9", optional = true, default-features = false }
-enarx-exec-wasmtime = { version = "0.4.0", path = "crates/exec-wasmtime", optional = true, default-features = false }
 
 # binary dependencies
 # FIXME: The enarx-exec-wasmtime-bin crate should be merged with enarx-exec-wasmtime again, see https://github.com/enarx/enarx/issues/1798
